Add files via upload
Overview: The main objective of this analysis is to determine which product to advertise to each customer, that is to determine specific products to advertise to specific customers at what locations. To achieve this, we will be answering the following Questions;

1.	What is the Correlation between sales and Income?
Using the Average income as X and Average Purchases for the last six months (Sales) as Y, there exists, the scatter plot shows that there is a positive relationship between Average income and sales. This means that as customers income increases, sales also increase. The Coefficient of determination (R^2) of 0.88 indicates that there is a strong relationship between both variables.
2.	What is the correlation Value between customer ratings and product return rate?
The analysis as seen in the scatter plot shows that there exists an inverse relationship between customer rating and return rate, this means that, as customer ratings increases, return rate decreases. The R^2 of 0.69 shows that there is a moderate relationship between both variables.
3.	What is the Linear regression formula to predict income from customer sales?
To predict the customer income from sales I used :X= b-y/-m where Customer Income as X and Average Purchases as Y. Predicted Income(X) = -722.14 - 'Customer List'[Last 6 Months Purchases]/-0.011
4.	Which customer do you predict has the highest income?
Using the Linear regression in question 3 above, the customer with the highest income as indicated by a funnel in the visualization is customer Jon with the customer ID JLit30836
5.	Which Product will be advertised the most?
From my analysis as shown in the table contained in the sales and income page, products should be advertised based on estimated income. 
Other Analysis:
•	I used the map to analyze top 4 states with the highest customer count. They include California, Florida, Texas and New York which are also the top four state with the highest population as indicated in the decomposition tree. However even though Texas has a higher population than Florida, analysis shows that Florida has more customers than Texas.
•	The cards in the income and sales page shows that there are only 1,000 customers out of a population of 159 million people across the 51 states. Which means that less than 1% of the population purchases our products.

Recommendations:
Based on the analysis above, below are my recommendations;
•	Advertisements should be based on estimated income of customers. There should be different campaigns for different income groups as shown in the product recommendation table
•	Marketing campaigns should be designed to attract states with lower population.
•	The more expensive products should be marketed more in states with high income levels while the less expensive products should be marketed in states with low-income level.
•	Efforts should be put into producing quality products in order to increase customers rating which will lead to a decrease in return rate and thus increase sales.